handle editing meta data
